+++ src/components/repeating-days-markup.js
@@ -1,0 +1,21 @@
+const createRepeatingDaysMarkup = (days, repeatingDays) => {
+  return days.map((day, index) => {
+    const isChecked = repeatingDays[day];
+
+    return (
+      `<input
+        class="visually-hidden card__repeat-day-input"
+        type="checkbox"
+        id="repeat-${day}-${index}"
+        name="repeat"
+        value="${day}"
+        ${isChecked ? `checked` : ``}
+      />
+      <label class="card__repeat-day" for="repeat-${day}-${index}"
+        >${day}</label
+      >`
+    );
+  }).join(`\n`);
+};
+
+export default createRepeatingDaysMarkup;
